chore: Improve the MetricType and MetricValue admin
Make it more useful as well as extract to its own file, the 1000+
base.py file is too large.
