+++ iaso/admin/__init__.py
@@ -2,6 +2,7 @@ from django.contrib import admin
 from django.contrib.auth.models import User
 
 from .base import IasoJSONEditorWidget  # noqa: F401
+from .metric_admin import *
 from .user_admin import UserAdmin
