+++ iaso/admin/base.py
@@ -56,8 +56,6 @@ from ..models import (
     Mapping,
     MappingVersion,
     MatchingAlgorithm,
-    MetricType,
-    MetricValue,
     OrgUnit,
     OrgUnitChangeRequest,
     OrgUnitChangeRequestConfiguration,
@@ -1254,31 +1252,6 @@ class DataSourceVersionsSynchronizationAdmin(admin.ModelAdmin):
                 "created_by",
             )
         )
-
-
-@admin.register(MetricType)
-class MetricTypeAdmin(admin.ModelAdmin):
-    list_display = (
-        "id",
-        "account",
-        "name",
-        "category",
-        "source",
-        "units",
-        "created_at",
-        "updated_at",
-    )
-    search_fields = ("name", "description", "source", "units", "comments")
-    list_filter = ("account", "source")
-    ordering = ("name",)
-
-
-@admin.register(MetricValue)
-class MetricValueAdmin(admin.ModelAdmin):
-    raw_id_fields = ("org_unit",)
-    list_display = ("metric_type", "org_unit", "year", "value")
-    search_fields = ("metric_type__name", "org_unit__name")
-    list_filter = ("metric_type", "year")
 
 
 admin.site.register(AccountFeatureFlag)
